Cleanup segmentation frontend code
- Reorganize field files into directories
- Extract woo fields/validation to match how subscriber fields work
- Fix typo in a file name
- Create new FilterProps type instead of repeating it all over the place

MAILPOET-5227
